LKE-12225 Query builder v1 (#759)
* LKE-12225 Feature flags config (#756)

* LKE-12225 Feature flag config

* LKE-12219 post review

* LKE-12088 data model for query builder (#758)


* LKE-12088 removing "readonly" modified on all models


* LKE-12088 support multiple node-categories or edge-types in filters (empty array means "any")


* LKE-12088 fixed type in query example comment


* LKE-12237 Convert structured queries to cypher (#760)

* LKE-12237 convert structured queries to cypher

* LKE-12237 update interfaces

* LKE-12237 export BasePropertyFilter

* LKE-12237 export interfaces and types for param validation

* Update queryBuilder.ts

* LKE-12237 added missing errors

* Update queryBuilder.ts

* Update queryBuilder.ts

* LKE-12412 Add queryBuilder feature flag (#767)

* LKE-12412 add queryBuilder feature flag

* LKE-12412 update feature name

---------
